Only log a warning if GRAALVM_HOME is not set and SSL is enabled
It was an issue prior to my SSL fixes but was hidden as
shamrock.ssl.native was never enabled on CI.

This is now the case as it can be enabled by default by an extension.

For the record, CI currently does not have GRAALVM_HOME set and uses a
Docker environment to build native images.