Change the way the SSL properties are defined
We want to allow the user to override these properties so we don't set
them unconditionally.

The idea would be to add additional paths to java.library.path but
there's something fishy going on with GraalVM: as long as there are more
than one paths in java.library.path, GraalVM can't find the library
anymore.