Improve reflection declarations for RESTEasy serialization
Ignore one more class and parameters marked with @Context.
Ignore the reference to the outer class: this hurts a lot when you have
an inner class in a REST resource not being static.
Improve the log message.

This is based on feedback from George Gastaldi on its migration to
Quarkus for the Fabric8 launcher backend project.