Springs Component is a Stereotype
This extends the spring integration to reflect that component
annotations are stereotypes. The ArC stereotype handling is extended to
allow providing additional annotations that are classed as stereotypes
along with a set of instances of those annotations that we want to
process (as some call sites may be unsuitable when processed using the
CDI logic).

The spring processor is enhanced to traverse all the annotations found
in the archive and testing them for the built in stereotypes and any
other stereotypes found previously in the traversal. From this we build
a graph of all the available stereotypes and their declared scopes, look
at the deployments beans and map the scopes and stereotypes to their CDI
equivalent.

Lastly, the CDI scope handling is tightened up to prevent duplicate
scopes being allowed on a bean.