Fix SSL on Mac OS X with GraalVM 1.0.0-rc14
Improved a bit the java.library.path detection. It's still pretty naive, but works.
The real issue is that substrateVM does not allow multiple element in the java.library.path, so we must pick one.
The detection is based on the number of files in the amd64 directory. However this is definitely brittle, and if the next versions of GraalVM adds file in this directory, we would need to increment the number.

This fixes #1665