Fix install when using uberjar
`mvn clean install` will currently fail when using uberjar,
unless the project is set to also install the sources or javadocs (which
is why the main build does not fail).

This changes the build so that the original artifact is always installed,
and when using an uberjar the uberjar is installed with the classifier 'runner'.