Swagger UI Docs
I had some trouble enabling the swagger ui and had to dive into the PR to figure out that there is a seperate extension that needed to be installed https://github.com/quarkusio/quarkus/pull/1573#issuecomment-478029052. I added some clarification to the docs detaling that 1) there is a separate extension for swagger-ui and 2) it is only available as of quarkus 0.14.0

Thanks for the good work folks, really enjoying building on quarkus.