Provides a reactive API and an implementation. Improve configuration and feature coverage in native mode.
Inject both reactive and imperative client
Extend configuration support for both the imperative and reactive clients
Add an integration test for the reactive client
Fix credential and write concern configuration
Remove the new for having quarkus-reactive-streams-operators as a dependency of projects using the mongo client
Re-inject the client so it get the new config
Close the client during the hot reload so we don't have a connection shortage
Register SSL support if not explicitly disabled
Provide a substitution for the DNS lookup. However it won't work until Substrate provide a non-blocking multicast implementation.
Declare the mongo version and the mongo extension in the bom.