Move the initialization of the delayed logger handler up in the RuntimeRunner
We used to have a workaround for dev mode but, in fact, we also have the
issue in the QuarkusTestExtension and, basically, in everything using
the RuntimeRunner.

Typically, for some test failures, you didn't have anything logged
because the delayed handler was never initialized.

So let's move the forced initialization of the logger in the
RuntimeRunner itself.