Create a shared API for local project workspaces
Adds a new `LocalProject` type to the bootstrap-core resolution package,
with the intention of modeling the existing Maven project workspaces
using it and also writing a Gradle backed implementation.

The responsibility of creating an AppModelResolver is relegated to the
LocalProject implementation, allowing the BoostrapClassLoaderFactory to
be configured with a build-system specific implementation as needed.