Create a loader for detecting and opening projects
Replaces existing local project discovery in BootstrapClassLoaderFactory
with a new LocalProjectLoader type that walks the ancestors of the
directory containing the build artifacts until a project descriptor file
is found (pom.xml at the moment).

Additionally, disable Maven project specific code in the
BootstrapClassLoaderFactory to allow retro-fitting and testing of a new
Gradle implementation.