Add Gradle implementation of LocalProject
Extends the bootstrap-core module with a Gradle LocalProject
implementation that resolves the AppModel by interacting with the
Quarkus Gradle plugin over the Gradle tooling API. This works by
exporting a set of interfaces representing the model that Gradle can
proxy and transfer over the wire.

Also extends the LocalProjectLoader to look for Gradle projects,
allowing tests annotated with `@QuarkusTest` to be run from Gradle
without a POM present.