Don't log a warning for quarkus.version property
Maven and Gradle properties are now pushed to config and if the property
name starts with quarkus. and the property is not consumed, we get a
warning.

This is especially the case for quarkus.version and it should obviously
be ignored.

This is just a workaround, I think we need a more general solution.