fix: avoid to fail the test when the database cannot be stopped
flapdoodle generates an error when it cannot stop the DB. Most of the time the DB will still be stopped (later) and as we cannot do anything about it let's not fail the test for it. 
Runnin test in a container ensures that if the DB cannot be stopped it will not prevent the test to run again in a separated container so it will not corrupt the build pipeline.