Change how filter are applied on routes
The previous approach was not applying on routes registered by the user directly on the Router object.
Also duplicate cors tests to the vertx-http extension so we test the new behavior, as well as with route using @Route

This commit also fixes how PEM certificate/key files are loaded.
It also adds tests for the HTTPS configuration of the Vert.x HTTP server.