Do better cleanup of tests if the JVM is killed
We write out all the created files to a file, and if this
file is still present we know that the JVM was killed and
we clean up all listed entries. Note that this can still
give problems for classes that use QuarkusTestExtension,
as JUnit 5 does not let us hook in before classloading.