Provide a way to set the TCCL in @BuildStep
This allows for the TCCL to automatically set the TCCL
in a build step. We can't do it for everyone as the metadata
needed to build this CL is created by @BuildStep methods.

Any build step that needs to load application classes should
use this flag to get a transformation safe ClassLoader.