Fixes Idea modules issue while importing the project.
In IDEA, when importing `devtools/gradle` as a Gradle project, in Project Structure, if you click to remove the `quarkus-gradle-plugin` project, and apply, the following message is displayed:

```
Content root "<WORKSPACE>/quarkus/bom/runtime" is defined for modules "quarkus-bom" and "quarkus-platform-descriptor-json".
Two modules in a project cannot share the same content root
```

This is because the `<resource>` entries in `platform-descriptor-json` are set to other projects, which may produce unpredictable results in IDEs.