Improve the Neo4j guide.
The Neo4j driver has been updated to 4.0.0 final and Neo4j 4.0 can be
used from Docker, the preview flag can now be removed.

The final driver revealed some errors in the existing guide:

The reading of the transactional result of a transactional function
needs to happen inside the same transaction.
Therefor the guide needs a fix so that the composition of the future
extracting the node happens in the transaction, not outside.