Improving the mailer adoc by adding quarkus.mailer.mock=false to sendgrid example
The sendgrid example in the current state will not work because the quarkus.mailer.mock=false is not present. If you just try to send the email it will fail and you will have to read the whole testing section to find out that the quarkus.mailer.mock is true by default.

I find this miss leading I personally have lost around 40mins finding and debugging why the email is not send and finding that I am in a mock mode.

I greatly believe that this will help the others using mailer to send email with or without sendgrid.

Maybe it is wise to even add a comment somewhere bellow the properties similar on how the password vault is explained.