Enforce the use of Maven 3.6.2 or later for the Quarkus build
Maven 3.6.1 and earlier does not seem to correctly resolve exclusions
and dependency management in dependencies.

The <exclusions> element in particular is heavily used in in
Gradle-generated POMs, and is apparently ignored by Maven when
computing transitive dependencies.

This can lead maven-enforcer-plugin to witness multiple dependencies
to the same artifact with different versions, which breaks the
dependency convergence rule and fails the build.

The issue is fixed in Maven 3.6.2, probably because of the upgrade
to Maven-Resolver 1.4 which fixes this seemingly related ticket:
https://issues.apache.org/jira/browse/MRESOLVER-33

In our case, the only affected transitive dependencies are declared by
the Elasticsearch REST client, which we depend on in the Elasticsearch
REST Client extension and in the Hibernate Search Elasticsearch
extension.