fix how defaultMode in Kubernetes volumes is parsed
Previously, the `defaultMode` config property of some Kubernetes
volumes (ConfigMap, secret) was always treated as a decimal number,
even though octal number is a common format (and we actually used
an octal number when specifying the default value!).

This commit changes the type of the config property to `String`,
which is later parsed to an `int` when calling the Dekorate API.
At that time, if the string begins with `0`, it's parsed as octal,
otherwise as decimal.