fix OpenTracing trace context propagation with Fault Tolerance @Asynchronous methods
There used to be a test for that, but didn't involve `@Asynchronous`
methods, so was in fact useless. This commit changes the test to
actually call an `@Asynchronous` method, which revealed an issue.
The `smallrye-fault-tolerance-tracing-propagation` dependency was
missing, so trace context propagation didn't actually work.

Then, the test also uncovered an issue in SmallRye Fault Tolerance
(see https://github.com/smallrye/smallrye-fault-tolerance/pull/208),
which is why this commit also brings a new version of it.

In addition to that, this commit also adds a test for tracing JDBC.