fix: Avoid logging a WARN message about having duplicates persistence-unit during testing
When running tests of an app with a `META-INF/persistence` file, I noticed the following
warning message:

```
2020-03-22 01:07:12,791 WARN  [org.hib.jpa.boo.int.PersistenceXmlParser] (build-22) HHH015018: Encountered multiple persistence-unit stanzas defining same name [templatePU]; persistence-unit names must be unique
```

The warning above is only observed in `test` profile and is deceiving since the app did not have mutliple persistence units.

Fix the issue by ensuring that no duplicates resources are loaded.

The error can be reproduced by simply running the tests (JVM mode) in the `ìntegration-tests/main` module.