fix(flyway): multiple datasource migration does not work on native
The QuarkusPathLocationScanner returned all migrations locations by default, this risked to cause
collision when we have a same filename for a migration script which is handled by different
datasources - a behaviour encountered on the issue #7685. Let's avoid this case by making sure that
the scanner only returns migration that it can recognize.

Fixes #7685