Parameterize the Java target
By default, it is based on the Java version used to create the project.
It can be overridden in CreateProject though if you really want to
enforce it in your tools.

I haven't exposed it for now as we will probably remove it or tweak it
later.

Note that even if you enforce it, it will still use either 8 or 11 for
now. We might want to further enhance that later.