Gradle: Refining the fix for #7996
The original fix for #7996 skipped a service provider if it showed up as part of the current classloader. This was done on the assumption that the parent classloader would contain the resource. But what happens if the parent classloader does not contain the reference? This commit attempts to refine the fix further by checking with the parent classloader, before banning the service provider implementation.