This PR fixes issues #8381 and #8382
    Unfortunately (or fortunately) it comes with a refactoring in the bootstrap maven resolver initialization.
Which also allowed me to add more tests to the TS (since previously a lot of the resolver parts where cached
using static variables and made it problematic).

    One of the behavioral changes (that fixes the issues mentioned above) is that now, if the workspace discovery
is enabled (which now is the default) the resolver initialization will be performed in the context of the current
project (if there is one, this would usually be relevant when running tests), which includes repository configurations.

    MavenRepoInitializer is now deprecated. Most of its logic has been copied to BootstrapMavenContext which is
initializing the relevant parts lazily whenever they are needed. MavenArtifactResolver which is the main resolver API
our build is using, is initialized from an instance of BootstrapMavenContext now. MavenArtifactResolver.Builder extends
BootstrapMavenContextConfig which exposes a lot more options, if necessary.