fix Jaeger metrics collection when SmallRye Metrics is absent
The Jaeger extension used to collect metrics unconditionally,
even when the SmallRye Metrics extension wasn't present.
With this commit, the Jaeger extension uses `NoopMetricsFactory`
if SmallRye Metrics are missing, and only uses the real
metrics factory if SmallRye Metrics are present.

This is not totally consistent with how metrics are enabled/disabled
in all other extensions. Fixing that is left for future.