fix Jaeger metrics initialization
Creation of the correct `MetricsFactory` is moved from runtime
to build time, which also means that it's created during native
image compilation. Therefore, the MicroProfile Metrics API and
SmallRye Metrics are not required to be present at runtime.
Which, in the end, means that native image compilation no longer
fails when the `quarkus-smallrye-opentracing` extension is present
but `quarkus-smallrye-metrics` is not.