Enable epic maker workflow
when committed then issues with label `kind/epic` will on
open/edit/labeling have
a section like:

```
<!-- EPIC:DATA
       #23 #445 #44
-->
```

Expanded to a task list with link to issue + title header.

the checkbox will reflect wether open or closed at the time of
edit of the issue.
(we can potentially improve that in future)

Includes update to epic template and disables debug printing as
to speed things up a bit.