Implement an annotation process to mark Panache archives
The logic here is that only libraries that are compiled
against panache will have this generated marker file, and
if this file is present this means that classes in the archive
may need to be transformed.

This is just a draft at the moment, a similar approach would
also need to be taken with Mongo.