Avoid build failures when cleaning up Gradle dev mode tests
We had several failures related to that with the build failing because
it was unable to delete a Gradle file. It's probably related to the
daemon but I couldn't find a reliable way to avoid starting the daemon
with the Gradle TestKit - I just found several people complaining about
it not working very well.

It's not a big deal if we let a couple of files survive the build from
time to time.