Speedup Flyway by preventing it from doing classpath scanning
This PR essentially takes the substitution we had for native mode
and turns it into a class transformer.
This way even in JVM mode our Flyway integration take advantage of
the fact that the migration script locations are known at build time,
thus no classpath scanning is needed.

This solution isn't ideal from a maintenance perspective, but the
transformation is relatively straightforward and our test suite for
Flyway pretty extensive, so we should easily be able to adapt to future
Flyway updates (ideally we would be able to remove this if /when
https://github.com/flyway/flyway/issues/2822 is done)

Relates to: #9428