Debug info improvements #9792
* Add third party source jars support.
* Third party jars not downloaded by default,
so it requires user to call `mvn dependencies:sources`
or equivalent.
* If source jars are found,
they're put next to the jars in the lib directory.
* After generating native executable,
the source jars are removed.
* Application sources are added individually using
-H:DebugInfoSourceSearchPath option.
They're only added if the directory exists.
* Debug symbols flag switched
from -H:GenerateDebugInfo=1 to `-g`.
* Sources cache has been moved to target/sources-cache
using the -H:DebugInfoSourceCacheRoot option.
This makes sources cache more closely available.
gdb should discover these automatically,
but it does not seem to work,
so better use `set directories` pointing to them.
* These changes only work on Mandrel or GraalVM 20.2,
so reworked the whole GraalVM version and distribution
resolution logic to make this easier to handle.
* Previous ad-hoc GraalVM version resolution
has been been replaced by a regex solution.
* Added unit tests to verify that
GraalVM versions and distributions are resolved correctly
with a diverse set of `native-image --version` outputs.