[#10295] Allow using the informer interface, and track down all types
This change also evaluates implementations of `ResourceEventHandler`,
which is the corresponding interface for the informer. This allows to
use watchers or informers the same way.

Also it traverses through the type information, adding all required
types to the list of types enabling reflection. If only the direct class
is added, deserialization will fail, unless all the references classes
are manually annotated with @RegisterForReflection.

Signed-off-by: Jens Reimann <jreimann@redhat.com>