Re-subscription must re-trigger the operation.
When an operation returns a Uni, resubscribing to this Uni must re-execute the underlying operation (except if the user explicitly uses cache()).
As RestEasy uses CompletionStage behind the scene, this behavior was not working and was always returning the same CompletionStage (the initial one). As a consequence, resubscriptions and retries were not working.

This commit changes how the Unis are created to re-execute the operation on every subscription. Note that it also delays the initial execution until the first subscription.