support reading configuration from Kubernetes Secrets
The `kubernetes` extension automatically generates a RoleBinding
that refers to the `view` ClusterRole. This ClusterRole doesn't
allow access to secrets. This commit therefore adds a configuration
property which, when enabled, makes the `kubernetes` extension
generate a special Role `view-secrets` and a second RoleBinding
referring to that role. This configuration property is build-time
only and has no other effect.

With this configuration in place, there's nothing preventing
the application from reading Secrets directly from the API server.

For convenience, a warning is printed at runtime if configuration
is read from Secrets yet the property is disabled.