Allow each module to have its own class path
Each module now contains a list of dependencies,
so individual class paths are respected.

Partially Fixes #10833, as this will now work when
quarkus:dev is run from the root, but not when it is
run from a sub folder. As we currently can't resolve
the dependencies of the other project.