Use the resolved paths to build the ClassLoader
This is a partial fix for #11144. At present
QuarkusTestExtension essentually guesses where
the test classes are, and these are used to build
the ClassLoader.

The Gradle resolver can add additional paths to the
AppArtifact (e.g. the testFixtures path), however they
will not actuall end up in the ClassLoader which causes
problems.