Don't skip the doc in the initial JDK 11 build
While we have a specific doc build now, we still need to build the doc
when the config classes have changed and having a pattern catching all
of them is relatively brittle.

Better build the doc anyway in the Initial JDK 11 build.

I thought about it because of the latest Hibernate Search upgrade that
touches the config classes but not the guide.