feat(redis): make the RedisConfiguration matches the upstream `RedisOptions`
This will make it easier to tweak things like redis connection hosts to match the expected endpoint
by the upstream library thus avoid manually reconstructing the connection string. See
https://github.com/vert-x3/vertx-redis-client/blob/e9e3c0db349d662e676c04235841c68488d6f6d2/src/main/java/io/vertx/redis/client/RedisOptions.java
for more details about the available options. The global default password for cluster and sentinel
mode will be added once we have the next release of vertx-redis. `NetClientOptions` can be added now
but I considered it to be of less importance and we can have separate PR to configure it once users
asks for it.

Fixes https://github.com/quarkusio/quarkus/issues/11986