Fix the handling of the port 0
By updating the options object, it create a race conditions.
One verticle may be already listening and would have updated the port in the HTTP option.
So, the next verticle see this set port (no more 0) and try to bind on that port, which lead to a BindException.

This commit stores the actual HTTP and HTTPS port without modifying the HTTP options.