Avoid copying sources directly in the source cache
This has the following benefits:
1. Quarkus code doesn't make any assumptions about the source-cache
structure or its name. Previously it assumed that the source-cache is
called sources and it contains a subfolder src with the applicatino
sources, which no longer holds for Mandrel 20.1.0.2 and won't hold for
GraalVM CE 20.3+.
2. The source-cache doesn't need to contain all the source files of the
application, only the ones being referenced by the debug info. As a
result, letting Graal construct the source cache we might end up with
less files in the source-cache.