Provide better DCE information when Agroal is not around
When Agroal is not around, we should never delegate to the JDBC-based
PersistenceProvider as things will fail, given we don't have a JDBC
datasource around.

Obviously, the test we introduce is not sufficient as having Agroal
around is not enough to have a proper Hibernate ORM setup but that's the
best we can do and should be sufficient in most cases.