Fix jvm.info metric naming to match micrometer convention
This patch brings the jvm.info metric in line with other jvm metric names and in line
with general micrometer naming conventions which define "." as the separator.
This will later automatically convert to "_" through a prometheus naming convention if prometheus
export is used. Even in a pure prometheus use-case the difference is relevantas metric mappers
and filters see the original name.