Add the option to perform hot reload via instrumentation
This does not require an app restart, so is much faster
for larger applications, but is limited to circumstances
where no structural changes have taken place.

This also adds the ability to check which classes have
been changed and what these changes actually are, which
allow frameworks to potentially cache metadata between
runs.

It also includes initial support for caching hibernate
proxies and transformations, as generating these are quite
expensive.