Vastly speed up META-INF/microprofile-config.properties for the common case
This file is always looked up during boot (as part of the spec) and most of the time
it doesn't exist.
Due to it's location under META-INF, it forces the RunnerClassLoader to look
into every jar order to determine whether or not it exists.
In order to avoid that lookup, we add a fail-fast path which is almost certainly
going to be used in most applications (which just use application.properties)