Make BytecodeRecorderImpl usable without reference to StartupTask
BytecodeRecorderImpl is an amazing piece feat of engineering.
Although it might not be obvious, it can be used as a general recording
mechanism, which is the point of this change.
By removing the hard coupling to StartupTask (), we could potentially generate
other classes as well.
This came about from some exploratory work on #13256.