Reactive routes - fix request context propagation
If there is a route filter and the request can have body (POST, PUT, etc.) then
the route method is invoked asynchronously (once all data are read). However,
the request context is activated by the filter and so we need to make sure the
same context is then used in the route method.

- resolves #13073