Issue with the SecurityIdentity on the IO Thread
This will no longer wrap the Identity in a Uni if
proactive auth is in use.

This also resolves a problem with setting the current
SecurityIdentity via a CDI event, which is not
compatible with lazy authentication.

Note that the problem can still occur if using lazy
auth from the IO thread, which is expected. To get
the identity in this situation you need to use
CurrentIdentityAssociation#getDeferredIdentity
and subscribe to the result.

Fixes #13835