Prevent UOE when MessageBodyReader can handle multiple media types
The basic idea of the PR is to get rid of the useless usages of MediaTypeHelper.getBestMatch.
Because of that we can now enforce the immutability of the media types in
ResourceReader and ResourceWriter.

Furthermore the PR applies some extra polish to MediaTypeHelper and
does away with the Iterator allocation in during the lookup
of the appropriate MessageBodyReader and MessageBodyWriter classes

Related to: https://github.com/quarkusio/quarkus/pull/13997#issuecomment-767260680