Include a list of dependencies in the app
This will allow vulnerability scanners to easily determine if
you application has any vulnerabilities.

Note that the 'vulnerability scanner' bit is still very much
a work in progress, however it makes more sense to include
this information on our side, then work on the scanner
once we have something it can work with.