Change the way the builder image is used on remote docker daemons
This patch creates a volume for /project where we can copy files to/from
using `docker cp`. The volume is create in the prebuild step, then used
in the build step, and finally removed in the postbuild step. The new
approach allows us to execute more than one commands using
`docker run --rm` on the same files in /project.