Introduce continuous testing
This commit adds the ability to test Quarkus applications
running in dev mode in a continous manner, only running
tests that are affected by the changed code.

To this is it changes the test ClassLoading model to work
the same as the dev mode model, and adds the ability to
launch a completly isolated test application inside the
same JVM as an existing Quarkus app.