Reactive DS idle-timeout for pooled connections
Before Vert.x 3.9.6, the reactive SQL client pool implementation did not support idle-timeout.

So we temporarily used the TCP connection idle-timeout config to achieve the same effect (with the caveat that the connection could be closed as idle after it was borrowed but not yet used for request execution).

With https://github.com/quarkusio/quarkus/pull/16100, Quarkus uses Vert.x 3.9.6 so we can now apply the idle-timeout config at the pool level and get rid of the warnings in the doc.

(cherry picked from commit 9310adc74840aade692db86da28e95dcbe78b9b1)