Manage docker-java version in the BOM
Testcontainers has a dependency on docker-java, and the two versions
must be in sync. If they are not, class loading errors occur.

Quarkus BOM manages Testcontainers, but does not manage docker-java.
(Note that Testcontainers BOM doesn't manage docker-java either.)

Which means that if user has a dependency that transitively brings in
both, such as `strimzi-test-container`, it may happen that incompatible
versions of Testcontainers (as managed by Quarkus BOM) and docker-java
(as managed by `strimzi-test-container`) are present, leading to
hard-to-diagnose problems.

Proper fix, in my opinion, is for Quarkus BOM to manage docker-java.
Unfortunately, docker-java doesn't provide a BOM, so we have to
spell out all its artifacts.