Update the default configuration of the `io.netty.allocator.maxOrder` property.
This property has a direct impact on the memory consumption of the application.
The current default is 1, while Netty itself uses 11.
While `1` worked well, recently, we started seeing issues, as Netty has assert statements failing with `1`.
Of course, you only see the issue if you run the application with assert enabled (in the IDE and some hot reload tests).

The extensions can request a higher value. The gRPC and RESTEasy Reactive extensions request `3`.

This commit updates the default value to `3` and provides a way for users to override the value.
If the user-configured value is lower than the extension request, it prints a warning during the build.
No matter what, the user-configured value is used.