Load base codestarts before extra codestarts
The precedence for the loaded codestarts should be `extensions < base < catalog < extra` so that the codestarts in the platform (i.e. catalog) and the extra provided codestarts can override any of the extension and base codestarts. This was not the case for base codestarts, which could not be overridden at all.