Make support for advanced Jackson features optional in RESTEasy Reactive
This is done by only engaging the MessageBodyWriter that supports these features
when the application actually makes use of the features.

The reason for doing this is to not pay the performance cost of checking
if each resource method makes use of these features if we know that
no resource method of the application uses them.