Update Vert.x to 4.1.0.Beta1 and Netty to 4.1.65.Final
This update introduces some changes.
First, when the HTTP compression is set to identity, the header is not written into the response anymore.
Then, the cookie interface has a new method (maxAge), so our implementations must be updated.

Finally, and that's the more significant change. Netty now depends on Brotli4J. Brotli is a compression algorithm.
The implementation uses a native dependency. At the moment, we won't support Brotli in native mode. This
commit adds a substitution to disable it. All decoders using Brotli related classes have been substituted to avoid having to depend on the Brotli4J dependency.