Ensure that VertxInputStream doesn't result in infinite attempts to read data
This change is needed because Netty now calls available to determine
whether or not more bytes are available, before calling read.
The negative value that we used to use does not play well with
PushbackInputStream which is used by Netty to wrap the stream.
For a full description of how I can came to this, see
https://github.com/quarkusio/quarkus/issues/17506#issuecomment-851365258

Fixes: #17506