Make VertxInputStream#available not fail on large content-type
An InputStream cannot read over 2GB of memory, but as `available`
is only an estimate, there is no good reason for that specific
method to fail if content-type indicated a larger payload.
Whether or not the processing fails or how it proceeds, should
be up to consumer of the stream and will likely have to do with
how `read` is implemented.

Fixes: #17455
(cherry picked from commit 017c53c3ffde1b18711426ecd6ceea63313073a3)